requirements.txt
ipykernel
pandas
matplotlib
seaborn
numpy
scikit-learn
geopandas
shapely
folium
statsmodels
nltk
tensorflow
category-encoders
xgboost
flask
ensure
python-box
pyyaml
dataclasses
streamlit
geopy


